+++ NEWS.md
@@ -1,9 +1,11 @@
 ## 20.2.0 IN-PROGRESS
 
 * Adds `effective location` attribute to holdings record (MODINVSTOR-669)
+* Adds `effective location` attribute to inventory-hierarchy API (MODINVSTOR-670) 
 * Defaults instance.previouslyHeld to false (MODINVSTOR-454)
 * Provides `holdings-storage 4.6` (MODINVSTOR-669
 * Provides `instance-storage 7.7`
+* Provides `inventory-hierarchy 0.2` (MODINVSTOR-670)
 
 ## 20.1.0 2021-03-25