+++ NEWS.md
@@ -1,6 +1,6 @@
 ## 20.2.0 IN-PROGRESS
 
-* Adds `effective location` attribute to holdings record and inventory-hierarchy API (MODINVSTOR-669)(MODINVSTOR-670) 
+* Adds `effective location` attribute to holdings record and inventory-hierarchy API (MODINVSTOR-669, MODINVSTOR-670) 
 * Defaults instance.previouslyHeld to false (MODINVSTOR-454)
 * Provides `holdings-storage 4.6` (MODINVSTOR-669
 * Provides `instance-storage 7.7`