+++ NEWS.md
@@ -1,3 +1,7 @@
+## 20.1.1 2021-04-22
+
+* Adds full-text index for classifications field in instance table (MODINVSTOR-716)
+
 ## 20.1.0 2021-03-25
 
 * Introduces `contributorsNames` CQL index for searching `instances` (MODINVSTOR-705)